Split extHost webview panels to own file
